Prometheus metrics and pyramid blocker

+++ internal/pkg/postgres/types.go
@@ -3,27 +3,27 @@ package postgres
 import "time"
 
 type User struct {
-	ID        int
+	ID        int64
 	Name      string
 	Flags     uint32
 	CreatedAt time.Time
 }
 
 type Quote struct {
-	ChannelID int
-	SenderID  int
+	ChannelID int64
+	SenderID  int64
 	Sender    string
 	Message   string
 	SentAt    time.Time
 }
 
 type Lines struct {
-	ChannelID int
-	UserID    int
-	Total     int
-	Unique    int
+	ChannelID int64
+	UserID    int64
+	Total     int64
+	Unique    int64
 	MostDate  string
-	MostCount int
+	MostCount int64
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
